Using TypeScript instead of JavaScript.
