+++ .gitignore
@@ -28,3 +28,7 @@ yarn-error.log*
 .env.development.local
 .env.test.local
 .env.production.local
+
+# Next.js files
+next-env.d.ts
+
