+++ .sequelizerc
@@ -1,7 +1,7 @@
 const path = require("path");
 
 module.exports = {
-    config: path.resolve("src", "db", "config", "config.js"),
+    config: path.resolve("src", "db", "config", "config.mjs"),
     "models-path": path.resolve("src", "db", "models"),
     "seeders-path": path.resolve("src", "db", "seeders"),
     "migrations-path": path.resolve("src", "db", "migrations"),
