+++ next.config.mjs
@@ -1,4 +1,5 @@
-module.exports = {
+/** @type {import('next').NextConfig} */
+const nextConfig = {
     webpack(config) {
         config.module.rules.push({
             test: /\.svg$/,
@@ -8,3 +9,5 @@ module.exports = {
         return config;
     },
 };
+
+export default nextConfig;
