+++ package.json
@@ -19,6 +19,7 @@
         "react": "18.2.0",
         "react-dom": "18.2.0",
         "sequelize": "^6.37.3",
+        "sequelize-typescript": "^2.1.6",
         "sqlite3": "^5.1.7"
     },
     "engines": {
@@ -26,6 +27,7 @@
     },
     "devDependencies": {
         "@svgr/webpack": "^8.1.0",
+        "@types/react": "18.3.12",
         "eslint": "8.57.0",
         "eslint-config-next": "14.2.5",
         "prettier": "3.3.3",
@@ -33,6 +35,7 @@
     },
     "imports": {
         "#root/*.js": "./*.js",
+        "#root/*.ts": "./*.ts",
         "#root/*.css": "./*.css",
         "#root/*.svg": "./*.svg"
     }
