+++ src/db/config/config.mjs
@@ -1,0 +1,10 @@
+export const options = {
+    dialect: "sqlite",
+    storage: "/tmp/wordmuncher.sqlite",
+};
+
+export default {
+    development: options,
+    test: options,
+    production: options,
+};
